BFS
Implementing a BFS that will go through each station and treats each unvisited one as a goal state. Eventually, it will have a driver function that derives a path using this recursive function.
